Move GetMembersTests
